Create  Price comparison extension
This function takes in a URL as an argument and uses the fetch API to retrieve the HTML of the webpage. It then uses the text() method to convert the response to a string, so that it can be parsed for the desired price data.

You can call this function and pass in the URL of the website you want to scrape, and it will return the extracted price data, you may need to extract the price data using regular expressions or a library like cheerio.